Update swagger types

+++ swagger_server/controllers/default_controller.py
@@ -14,7 +14,7 @@ def add_student(body=None):  # noqa: E501
     :param body: Student item to add
     :type body: dict | bytes
 
-    :rtype: float
+    :rtype: int
     """
     if connexion.request.is_json:
         body = Student.from_dict(connexion.request.get_json())  # noqa: E501
@@ -28,7 +28,7 @@ def delete_student(student_id):  # noqa: E501
     delete a single student # noqa: E501
 
     :param student_id: the uid
-    :type student_id: float
+    :type student_id: int
 
     :rtype: None
     """
@@ -38,10 +38,10 @@ def delete_student(student_id):  # noqa: E501
 def get_student_by_id(student_id):  # noqa: E501
     """gets student
 
-    Returns a single student  # noqa: E501
+    Returns a single student # noqa: E501
 
     :param student_id: the uid
-    :type student_id: float
+    :type student_id: int
 
     :rtype: Student
     """
